+++ lume/serializers/hdf5.py
@@ -1,4 +1,4 @@
-from .base import SerializerBase
+from .base import SerializerBase, ModuleImportError, ClassInitError
 import h5py
 import importlib
 import numpy as np
@@ -10,22 +10,6 @@ if TYPE_CHECKING:
     from lume.base import Base
 
 
-class ModuleImportError(Exception):
-    def __init__(self, module_name, module_version):
-        self.module_name = module_name
-        module_base = module_name.split(".")[0]
-        module_version = module_version
-        self.message = f"Unable to import module {module_name}. Object was serialized with {module_version}. Is a compatible version of {module_base} installed?"
-        super().__init__(self.message)
-
-
-class ClassInitError(Exception):
-    def __init__(self, class_name, module_name, module_version):
-        self.module_name = module_name
-        self.class_name = class_name
-        module_base = module_name.split(".")[0]
-        self.message = f"Unable to get {class_name} from {module_name}. Object was serialized with {module_version}. Is a compatible version of {module_base} installed?"
-        super().__init__(self.message)
 
 
 class HDF5Serializer(SerializerBase):
